requirements.txt
torch
numpy
librosa
pillow
transformers
gradio
soundfile
